requirements/dev.txt
-r prod.txt

ipython~=8.29.0
django-debug-toolbar~=5.0.0a0
ruff~=0.7.4
black~=24.10.0
ngrok
